requirements.txt
# Skynette - AI-Native Workflow Automation Platform
# Install with: pip install -r requirements.txt

# UI Framework
flet>=0.25.0

# AI & ML
llama-cpp-python>=0.2.0
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
groq>=0.5.0
huggingface-hub>=0.20.0
chromadb>=0.4.0

# Data & Storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pyyaml>=6.0
cryptography>=41.0.0
keyring>=24.0.0

# HTTP & Networking
httpx>=0.27.0
websockets>=12.0
fastapi>=0.110.0
uvicorn>=0.27.0

# Workflow Engine
apscheduler>=3.10.0
watchdog>=4.0.0
jinja2>=3.1.0
jsonschema>=4.21.0

# File Processing
pandas>=2.0.0
openpyxl>=3.1.0
pypdf>=4.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
aiofiles>=23.0.0

# Cloud Services
boto3>=1.28.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Build Tools
pyinstaller>=6.0.0

# Development (optional)
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0
# ruff>=0.3.0
